Add files via upload
The code provided searches for a keyword in the body of the latest specified number of emails in a Gmail inbox. It then counts the number of occurrences of the keyword and extracts the unique email addresses of the senders of the emails that contain the keyword. These unique email addresses are then exported to an Excel file named "sender_emails.xlsx".